test: Move to separate file
